requirements.txt
boto3
keras
minio
tensorflow
Pillow
opencv-python-headless